requirements.txt
Flask
flask-cors
chromadb
google-generativeai
python-dotenv
gunicorn